tests/wpt/tests/workers/interfaces/WorkerUtils/WindowTimers/001.html

<!doctype html>
<title>setTimeout</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('001.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 1);
    this.done();
  });
});
</script>
